cleaned the cpu kronmult
