+++ src/asgard_solver.hpp
@@ -1,5 +1,5 @@
 #pragma once
-#include "./device/asgard_kronmult.hpp"
+#include "asgard_kronmult.hpp"
 #include "asgard_pde.hpp"
 #include "asgard_pde_functions.hpp"
 #include "asgard_wavelet_basis.hpp"
